# contract/openapi.json
{
  "components": {
    "schemas": {
...
      "Activity": {
        "enum": [
          "running",
          "walking",
          "cycling",
          "swimming",
          "strength_training",
          "cardio",
          "stretching_mobility",
          "other"
        ],
        "title": "Activity",
        "type": "string"
      },
...
      "CompletionCreate": {
        "additionalProperties": false,
        "properties": {
          "activity": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Activity"
              },
              {
                "type": "null"
              }
            ]
          },
          "note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          },
          "on_date": {
            "format": "date",
            "title": "On Date",
            "type": "string"
          },
          "schedule_entry_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Schedule Entry Id"
          },
          "unit": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Unit"
              },
              {
                "type": "null"
              }
            ]
          },
          "value": {
            "anyOf": [
              {
                "exclusiveMinimum": 0.0,
                "type": "number"
              },
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Value"
          }
        },
        "required": [
          "on_date"
        ],
        "title": "CompletionCreate",
        "type": "object"
      },
      "CompletionRead": {
        "properties": {
          "activity": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Activity"
              },
              {
                "type": "null"
              }
            ]
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "label": {
            "title": "Label",
            "type": "string"
          },
          "note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          },
          "on_date": {
            "format": "date",
            "title": "On Date",
            "type": "string"
          },
          "schedule_entry_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Schedule Entry Id"
          },
          "source": {
            "$ref": "#/components/schemas/CompletionSource"
          },
          "unit": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Unit"
              },
              {
                "type": "null"
              }
            ]
          },
          "value": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Value"
          }
        },
        "required": [
          "id",
          "activity",
          "value",
          "unit",
          "on_date",
          "schedule_entry_id",
          "source",
          "label",
          "note",
          "created_at"
        ],
        "title": "CompletionRead",
        "type": "object"
      },
      "CompletionSource": {
        "description": "Written once at creation, never derived or updated - schedule_entry_id\ngets nulled out (by ON DELETE SET NULL) the moment its entry or workout is\ndeleted, so it cannot answer \"was this scheduled?\" after the fact. Plan\nadherence must not change retroactively just because someone tidied up\ntheir plan.",
        "enum": [
          "scheduled",
          "standalone"
        ],
        "title": "CompletionSource",
        "type": "string"
      },
      "CompletionUpdate": {
        "additionalProperties": false,
        "properties": {
          "note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          },
          "unit": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Unit"
              },
              {
                "type": "null"
              }
            ]
          },
          "value": {
            "anyOf": [
              {
                "exclusiveMinimum": 0.0,
                "type": "number"
              },
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Value"
          }
        },
        "title": "CompletionUpdate",
        "type": "object"
      },
      "DayScheduleRead": {
        "description": "No `date` field - the date is already the key this value sits under in\nScheduleResponse.days, so it isn't duplicated here.",
        "properties": {
          "cancelled": {
            "items": {
              "$ref": "#/components/schemas/EntryRefRead"
            },
            "title": "Cancelled",
            "type": "array"
          },
          "completed": {
            "title": "Completed",
            "type": "boolean"
          },
          "entries": {
            "items": {
              "$ref": "#/components/schemas/ResolvedEntryRead"
            },
            "title": "Entries",
            "type": "array"
          },
          "status": {
            "$ref": "#/components/schemas/DayStatus"
          }
        },
        "required": [
          "status",
          "entries",
          "cancelled",
          "completed"
        ],
        "title": "DayScheduleRead",
        "type": "object"
      },
      "DayStatus": {
        "description": "Day-level status is a presentation convenience for compact rendering\n(e.g. a single calendar-cell badge) - `entries` and `cancelled` are the\nauthoritative data. Clients must not branch on this status alone in a\ndetailed view: a SCHEDULED day can still carry unrelated cancelled\nentries (see resolve()'s mixed-day precedence), and SUBSTITUTED only\nmeans *at least one* surviving entry that day is a substitution.\n\nThese string values are part of the public wire contract - app/schemas/\nimports this enum directly for the API response, so renaming a member\nhere changes the API, not just an internal detail.",
        "enum": [
          "empty",
          "scheduled",
          "cancelled",
          "substituted"
        ],
        "title": "DayStatus",
        "type": "string"
      },
...
      "EntryRefRead": {
        "description": "Minimal reference to another entry - used both for what a substitution\nreplaced and what a cancellation cancelled, since both need exactly the\nsame information: which entry, and what to call it. Never None.entry_id;\nthe whole reference is Optional at the call site instead, for the cases\nwhere there is nothing to reference at all.",
        "properties": {
          "entry_id": {
            "format": "uuid",
            "title": "Entry Id",
            "type": "string"
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          }
        },
        "required": [
          "entry_id",
          "name"
        ],
        "title": "EntryRefRead",
        "type": "object"
      },
      "EntryStatus": {
        "enum": [
          "scheduled",
          "substituted"
        ],
        "title": "EntryStatus",
        "type": "string"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ResolvedEntryRead": {
        "properties": {
          "completion_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completion Id"
          },
          "entry_id": {
            "format": "uuid",
            "title": "Entry Id",
            "type": "string"
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "replaced": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/EntryRefRead"
              },
              {
                "type": "null"
              }
            ]
          },
          "status": {
            "$ref": "#/components/schemas/EntryStatus"
          },
          "workout_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Workout Id"
          }
        },
        "required": [
          "entry_id",
          "workout_id",
          "name",
          "notes",
          "status",
          "replaced",
          "completion_id"
        ],
        "title": "ResolvedEntryRead",
        "type": "object"
      },
...
      "ScheduleResponse": {
        "properties": {
          "days": {
            "additionalProperties": {
              "$ref": "#/components/schemas/DayScheduleRead"
            },
            "title": "Days",
            "type": "object"
          }
        },
        "required": [
          "days"
        ],
        "title": "ScheduleResponse",
        "type": "object"
      },
      "Unit": {
        "enum": [
          "minutes",
          "hours",
          "miles",
          "kilometers",
          "sets",
          "reps"
        ],
        "title": "Unit",
        "type": "string"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "follow-through API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/completions": {
      "get": {
        "operationId": "list_completions_completions_get",
        "parameters": [
          {
            "in": "query",
            "name": "from",
            "required": true,
            "schema": {
              "format": "date",
              "title": "From",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "to",
            "required": true,
            "schema": {
              "format": "date",
              "title": "To",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "activity",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/Activity"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Activity"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/CompletionRead"
                  },
                  "title": "Response List Completions Completions Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Completions",
        "tags": [
          "completions"
        ]
      },
      "post": {
        "operationId": "create_completion_completions_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CompletionCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CompletionRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create Completion",
        "tags": [
          "completions"
        ]
      }
    },
    "/completions/{completion_id}": {
      "delete": {
        "operationId": "delete_completion_completions__completion_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "completion_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Completion Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Delete Completion",
        "tags": [
          "completions"
        ]
      },
      "patch": {
        "operationId": "update_completion_completions__completion_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "completion_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Completion Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CompletionUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CompletionRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Update Completion",
        "tags": [
          "completions"
        ]
      }
    },
...
    "/plans/{plan_id}/schedule": {
      "get": {
        "operationId": "get_schedule_plans__plan_id__schedule_get",
        "parameters": [
          {
            "in": "path",
            "name": "plan_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Plan Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "from",
            "required": true,
            "schema": {
              "format": "date",
              "title": "From",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "to",
            "required": true,
            "schema": {
              "format": "date",
              "title": "To",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScheduleResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Schedule",
        "tags": [
          "schedule"
        ]
      }
    },
...
  }
}